completando tareas

+++ Programacion_TP3/src/main/java/dao/EmpleadoDao.java
@@ -3,7 +3,9 @@ import entidad.Empleado;
 
 public interface EmpleadoDao extends Dao<Empleado>{
 
-	Empleado empleadoAutenticado(String contrasenia, Integer legajo);
+	
+
+	Empleado empleadoAutenticado(String contrasenia, Long legajo);
